refactor: collapse EventCollector + ServiceCollector to native SSR
Drop the SsrEventPanelCollector / SsrServicePanelCollector mirror classes
and have the actual `EventCollector` / `ServiceCollector` implement
`HtmlViewProviderInterface` directly: each collector now ships its own
template via `getViewPath()` and the API renders it on demand. One
collector → one panel, no parallel mirror class.

To make Kernel collectors implement the SSR contract, both
`HtmlViewProviderInterface` and `Slot` are promoted from the API module
to the Kernel module:

  libs/API/src/Debug/HtmlViewProviderInterface.php
    → libs/Kernel/src/Collector/HtmlViewProviderInterface.php
  libs/API/src/Debug/Slot/Slot.php
    → libs/Kernel/src/Slot/Slot.php
  libs/API/tests/Unit/Debug/Slot/SlotTest.php
    → libs/Kernel/tests/Unit/Slot/SlotTest.php

The slot contract (data-adp-slot, data-adp-payload) is unchanged — only
the PHP namespaces moved.

Adapter wiring: every Ssr*PanelCollector registration in Yii 3, Symfony,
Laravel, and Yii 2 is removed; the existing EventCollector /
ServiceCollector registrations now serve as both the data collector and
the SSR view provider. Spiral was untouched (already binds the bare
collectors).

Frontend: the React EventPanel + ServicesPanel components and their
Layout.tsx mappings are removed. Both collectors now flow through the
default `__html` branch which mounts SsrPanel — same hydration path as
any future SSR collector.

Net diff: 918 lines deleted (two custom React panels + four mirror
classes) vs. 30 lines added (the two new view templates + interface
implementation lines on each collector).

Tests:
  - PHP: API/Kernel 80/80; adapters 944/944; Spiral 81/81
  - Frontend (Debug module): 320/320
  - Modulite: 0 violations

End-to-end on Yii 3 playground:
  - GET /debug/api/view/<id>?collector=AppDevPanel\Kernel\Collector\EventCollector
    returns {__html: ...} with 25 events, 25 hydrated class-name slots,
    25 hydrated json slots
  - ServiceCollector: 62 service calls render through SsrPanel with full
    ClassName + JsonRenderer hydration; failed calls flip the chip to
    error severity via data-severity.
